Fix - Performance issue Add - Physics Interpolation

+++ CSC8503/PaintableObject.cpp
@@ -15,7 +15,7 @@ void PaintableObject::Update(float dt) {
 
 
 void PaintableObject::AddImpactPoint(ImpactPoint point) { //Impact point with world space cooridnates passed in; convert to local space.
-	int impactCount = impactPoints.size();
+	int impactCount = (int)impactPoints.size();
 
 	if (impactCount < 10) {
 		impactPoints.push_back(point);
